// src/utils/types/client.ts
export enum Status {
  CONTACTED = "Contacted",
  NOT_CONTACTED = "Not contacted",
  CONTACT_IN_FUTURE = "Contact in future",
}

export type Client = {
  id: string;
  firstname: string;
  lastname: string;
  email: string;
  phoneNumber: string;
  notes: string;
  status: Status;
};

export type Clients = Client[];

export type clientDto = {
  id: string;
  firstname: string;
  lastname: string;
  email: string;
  phoneNumber: string;
};